+++ custom_components/maintenance_supporter/frontend/strategy/chunks/chunk-JUQFJUGX.js
@@ -1,5 +1,5 @@
-/*! maintenance_supporter frontend 2.47.0 */
-import{a as r}from"./chunk-WQTIH66Y.js";var a=r`
+/*! maintenance_supporter frontend 2.48.0 */
+import{a as r}from"./chunk-NO26YKIV.js";var a=r`
   ha-card { overflow: hidden; }
   .card-content {
     padding: 16px;